Flag fixes for Europe map and for Brittany in flag menu and Gateway To the Atlantic map (#910)
## Description:

As reported here
https://discord.com/channels/1284581928254701718/1377035663404433490

Flag names in Europe.json didn't correspond with those in Countries.json
for: England, Scotland, Wales. So the Nations' flags weren't displayed
in-game. Changed them in Europe.json.

The flag of Brittanny was looked up as 'britanny' by both the flag menu
and GatewayToTheAtlantic.json. Changed its name in Countries.json to
correspond with this so the flag shows up in flag menu and in-game.

## Please complete the following:

- [ ] I have added screenshots for all UI updates
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I understand that submitting code with bugs that could have been
caught through manual testing blocks releases and new features for all
contributors

## Please put your Discord username so you can be contacted if a bug or
regression is found:

tryout33