Flag fixes in several maps (#957)
## Description:

Thanks to J.Dimlight for reporting this.

Flag images were not displayed on maps. Or the wrong flag was displayed.
Because of no flag code, or pointing to non-existant or incorrect flag
file names.

There are currently no flags to display for these Nations, as there is
no flag file present in resources:

Gateway to the Atlantic
- Jesuit Monks
- Wine

Japan and Neighbors:
- Sakhalin

Mars:
- European Space Agency

South America:
- The Biggest Snakes
- Normal Capybaras
- Just Otters

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I understand that submitting code with bugs that could have been
caught through manual testing blocks releases and new features for all
contributors

## Please put your Discord username so you can be contacted if a bug or
regression is found:

tryout33